requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
cloudpickle==3.0.0
    # via
    #   gym
    #   tensordict
    #   torchrl
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
filelock==3.13.1
    # via torch
fonttools==4.45.0
    # via matplotlib
fsspec==2023.10.0
    # via torch
gym==0.26.2
    # via nes-py
gym-notices==0.0.8
    # via gym
gym-super-mario-bros==7.4.0
    # via -r requirements.in
idna==3.4
    # via requests
imageio==2.33.0
    # via scikit-image
jinja2==3.1.2
    # via torch
kiwisolver==1.4.5
    # via matplotlib
lazy-loader==0.3
    # via scikit-image
markupsafe==2.1.3
    # via jinja2
matplotlib==3.8.2
    # via -r requirements.in
mpmath==1.3.0
    # via sympy
mypy==1.7.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via mypy
nes-py==8.2.1
    # via gym-super-mario-bros
networkx==3.2.1
    # via
    #   scikit-image
    #   torch
numpy==1.26.2
    # via
    #   contourpy
    #   gym
    #   imageio
    #   matplotlib
    #   nes-py
    #   opencv-python
    #   scikit-image
    #   scipy
    #   tensordict
    #   tifffile
    #   torchrl
    #   torchvision
opencv-python==4.8.1.78
    # via -r requirements.in
packaging==23.2
    # via
    #   matplotlib
    #   scikit-image
    #   torchrl
pillow==10.1.0
    # via
    #   imageio
    #   matplotlib
    #   scikit-image
    #   torchvision
pyglet==1.5.21
    # via nes-py
pyparsing==3.1.1
    # via matplotlib
python-dateutil==2.8.2
    # via matplotlib
python-dotenv==0.21.0
    # via -r requirements.in
requests==2.31.0
    # via torchvision
ruff==0.1.5
    # via -r requirements.in
scikit-image==0.22.0
    # via -r requirements.in
scipy==1.11.4
    # via scikit-image
six==1.16.0
    # via python-dateutil
sympy==1.12
    # via torch
tensordict==0.2.1
    # via
    #   -r requirements.in
    #   torchrl
tifffile==2023.9.26
    # via scikit-image
torch==2.1.1
    # via
    #   -r requirements.in
    #   tensordict
    #   torchrl
    #   torchvision
torchrl==0.2.1
    # via -r requirements.in
torchvision==0.16.1
    # via -r requirements.in
tqdm==4.66.1
    # via nes-py
typing-extensions==4.8.0
    # via
    #   mypy
    #   torch
urllib3==2.1.0
    # via requests
